# weather/service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a weather service RestAPI.",
        "title": "RestAPI",
        "contact": {
            "name": "dingowd",
            "email": "directtosun@mail.ru"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/cities": {
            "get": {
                "description": "Получить список городов",
                "produces": [
                    "application/json"
                ],
                "summary": "Список городов",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.City"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/detail": {
            "get": {
                "description": "Получить детальный прогноз",
                "produces": [
                    "application/json"
                ],
                "summary": "Детальный прогноз",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название города",
                        "name": "city",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата прогноза",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Resp"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/detail_favor": {
            "get": {
                "description": "Получить детальный прогноз избранных городов",
                "produces": [
                    "application/json"
                ],
                "summary": "Детальный прогноз",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя пользователя",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Дата прогноза",
                        "name": "date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Resp"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/insert_fav": {
            "post": {
                "description": "Добавить город в избранное",
                "produces": [
                    "application/json"
                ],
                "summary": "Добавить избранное",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя пользователя",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Название нового города",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/insert_user": {
            "post": {
                "description": "Добавить нового пользователя",
                "produces": [
                    "application/json"
                ],
                "summary": "Добавить пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя нового пользователя",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Res"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/short": {
            "get": {
                "description": "Получить краткий прогноз",
                "produces": [
                    "application/json"
                ],
                "summary": "Краткий прогноз",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Название города",
                        "name": "city",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ShortWeather"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        },
        "/short_favor": {
            "get": {
                "description": "Получить краткий прогноз избранных городов",
                "produces": [
                    "application/json"
                ],
                "summary": "Краткий прогноз",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя пользователя",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ShortWeather"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.Err"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.City": {
            "type": "object",
            "properties": {
                "country": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                }
            }
        },
        "models.CityProp": {
            "type": "object",
            "properties": {
                "coord": {
                    "$ref": "#/definitions/models.Coord"
                },
                "country": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "population": {
                    "type": "integer"
                },
                "sunrise": {
                    "type": "integer"
                },
                "sunset": {
                    "type": "integer"
                },
                "timezone": {
                    "type": "integer"
                }
            }
        },
        "models.Clouds": {
            "type": "object",
            "properties": {
                "all": {
                    "type": "integer"
                }
            }
        },
        "models.Coord": {
            "type": "object",
            "properties": {
                "lat": {
                    "type": "number"
                },
                "lon": {
                    "type": "number"
                }
            }
        },
        "models.List": {
            "type": "object",
            "properties": {
                "clouds": {
                    "$ref": "#/definitions/models.Clouds"
                },
                "dt": {
                    "type": "integer"
                },
                "dt_txt": {
                    "type": "string"
                },
                "main": {
                    "$ref": "#/definitions/models.Main"
                },
                "pop": {
                    "type": "integer"
                },
                "sys": {
                    "$ref": "#/definitions/models.Sys"
                },
                "visibility": {
                    "type": "integer"
                },
                "weather": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Weather"
                    }
                },
                "wind": {
                    "$ref": "#/definitions/models.Wind"
                }
            }
        },
        "models.Main": {
            "type": "object",
            "properties": {
                "feels_like": {
                    "type": "number"
                },
                "grnd_level": {
                    "type": "integer"
                },
                "humidity": {
                    "type": "integer"
                },
                "pressure": {
                    "type": "integer"
                },
                "sea_level": {
                    "type": "integer"
                },
                "temp": {
                    "type": "number"
                },
                "temp_kf": {
                    "type": "integer"
                },
                "temp_max": {
                    "type": "number"
                },
                "temp_min": {
                    "type": "number"
                }
            }
        },
        "models.Resp": {
            "type": "object",
            "properties": {
                "city": {
                    "$ref": "#/definitions/models.CityProp"
                },
                "cnt": {
                    "type": "integer"
                },
                "cod": {
                    "type": "string"
                },
                "list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.List"
                    }
                },
                "message": {
                    "type": "integer"
                }
            }
        },
        "models.ShortWeather": {
            "type": "object",
            "properties": {
                "av_temp": {
                    "type": "number"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "dates": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "models.Sys": {
            "type": "object",
            "properties": {
                "pod": {
                    "type": "string"
                }
            }
        },
        "models.Weather": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "main": {
                    "type": "string"
                }
            }
        },
        "models.Wind": {
            "type": "object",
            "properties": {
                "deg": {
                    "type": "integer"
                },
                "gust": {
                    "type": "number"
                },
                "speed": {
                    "type": "number"
                }
            }
        },
        "utils.Err": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "utils.Res": {
            "type": "object",
            "properties": {
                "result": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}